+++ src/components/pages/Imprint/index.tsx
@@ -18,7 +18,12 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { PageHeader, Typography } from '@catena-x/portal-shared-components'
+import {
+  Input,
+  PageHeader,
+  Typography,
+} from '@catena-x/portal-shared-components'
+import { Box } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 
 export default function Imprint() {
@@ -28,28 +33,124 @@ export default function Imprint() {
       <PageHeader headerHeight={200} topPage={true} title={t('title')} />
       <section>
         <Typography variant="h5">{t('directors')}</Typography>
-        <Typography variant="body2">Oliver Ganser ({t('ceo')})</Typography>
-        <Typography variant="body2">
-          Prof. Dr.-Ing. Boris Otto ({t('deputyCeo')})
-        </Typography>
-        <Typography variant="body2">
-          Claus Cremers ({t('treasurer')})
-        </Typography>
+
+        <Box sx={{ display: 'inline-flex' }}>
+          <div style={{ paddingTop: '20px' }}>CEO : </div>
+          <div style={{ marginTop: '-29px' }}>
+            <Input
+              sx={{
+                width: 'max-width',
+                ml: 1,
+                p: 0,
+                display: 'inline-flex',
+                '.MuiFilledInput-input': { padding: '10px !important' },
+              }}
+              placeholder={'Oliver Ganser'}
+              disabled={true}
+            />
+          </div>
+        </Box>
+        <div>
+          <Box sx={{ display: 'inline-flex' }}>
+            <div style={{ paddingTop: '20px' }}>{t('deputyCeo')} : </div>
+            <div style={{ marginTop: '-25px' }}>
+              <Input
+                sx={{
+                  width: 'max-width',
+                  ml: 1,
+                  p: 0,
+                  display: 'inline-flex',
+                  '.MuiFilledInput-input': { padding: '10px !important' },
+                }}
+                placeholder={'Prof. Dr.-Ing. Boris Otto'}
+                disabled={true}
+              />
+            </div>
+          </Box>
+        </div>
+        <div>
+          <Box sx={{ display: 'inline-flex' }}>
+            <div style={{ paddingTop: '20px' }}>{t('treasurer')} : </div>
+            <div style={{ marginTop: '-20px' }}>
+              <Input
+                sx={{
+                  width: 'max-width',
+                  ml: 1,
+                  p: 0,
+                  display: 'inline-flex',
+                  '.MuiFilledInput-input': { padding: '10px !important' },
+                }}
+                placeholder={'Claus Cremers'}
+                disabled={true}
+              />
+            </div>
+          </Box>
+        </div>
         <br />
         <Typography variant="body2">{t('address')}</Typography>
         <Typography variant="body2">
           Catena-X Automotive Network e.V.
         </Typography>
-        <Typography variant="body2">c/o IFOK GmbH</Typography>
-        <Typography variant="body2">Reinhardtstraße 58</Typography>
-        <Typography variant="body2">10117 Berlin</Typography>
+        <div>
+          <Box sx={{ display: 'inline-flex' }}>
+            <div style={{ paddingTop: '20px' }}>c/o : </div>
+            <div style={{ marginTop: '-20px' }}>
+              <Input
+                sx={{
+                  width: 'max-width',
+                  ml: 1,
+                  p: 0,
+                  display: 'inline-flex',
+                  '.MuiFilledInput-input': { padding: '10px !important' },
+                }}
+                placeholder={'IFOK GmbH'}
+                disabled={true}
+              />
+            </div>
+          </Box>
+        </div>
+
+        <div>
+          <Box sx={{ display: 'inline-flex' }}>
+            <div style={{ marginTop: '-20px' }}>
+              <Input
+                sx={{
+                  width: 'max-width',
+                  ml: 1,
+                  p: 0,
+                  display: 'inline-flex',
+                  '.MuiFilledInput-input': { padding: '10px !important' },
+                }}
+                placeholder={'Reinhardtstraße 58'}
+                disabled={true}
+              />
+            </div>
+          </Box>
+        </div>
+        <div>
+          <Box sx={{ display: 'inline-flex' }}>
+            <div style={{ marginTop: '-20px' }}>
+              <Input
+                sx={{
+                  width: 'max-width',
+                  ml: 1,
+                  p: 0,
+                  display: 'inline-flex',
+                  '.MuiFilledInput-input': { padding: '10px !important' },
+                }}
+                placeholder={'10117 Berlin'}
+                disabled={true}
+              />
+            </div>
+          </Box>
+        </div>
         <br />
         <Typography variant="h5">{t('contact&support')}</Typography>
         <Typography variant="body2">{t('contact&supportDesc')}</Typography>
         <br />
         <Typography variant="h5">{t('privacy')}</Typography>
         <Typography variant="body2">{t('privacyDesc')}</Typography>
-        <a href=".">xxx.xxx.xxx</a>
+        <a href="./privacy">Privacy Policy</a>
       </section>
     </main>
   )